+++ debug/debug.h
@@ -4,6 +4,6 @@
 #include "bytecode.h"
 
 void disassembleBytecode(Bytecode* bytecode, const char* name);
-int32_t disassembleInstruction(Bytecode* bytecode, uint32_t offset);
+int32_t disassembleInstruction(Bytecode* bytecode, uint16_t offset);
 
 #endif // !FUNVM_DEBUG_H
